+++ noodles-bam/src/record/codec.rs
@@ -1,3 +1,4 @@
 pub mod decoder;
+pub mod encoder;
 
-pub(crate) use self::decoder::decode_record as decode;
+pub(crate) use self::{decoder::decode_record as decode, encoder::encode_record as encode};