+++ noodles-bam/src/writer.rs
@@ -2,7 +2,6 @@
 
 mod builder;
 mod header;
-pub mod record;
 
 pub use self::builder::Builder;
 
@@ -11,8 +10,6 @@ use std::io::{self, Write};
 use byteorder::{LittleEndian, WriteBytesExt};
 use noodles_bgzf as bgzf;
 use noodles_sam::{self as sam, alignment::Record};
-
-use self::record::encode_record;
 
 /// A BAM writer.
 ///
@@ -122,8 +119,10 @@ where
     /// # Ok::<(), io::Error>(())
     /// ```
     pub fn write_record(&mut self, header: &sam::Header, record: &Record) -> io::Result<()> {
+        use super::record::codec::encode;
+
         self.buf.clear();
-        encode_record(&mut self.buf, header, record)?;
+        encode(&mut self.buf, header, record)?;
 
         let block_size = u32::try_from(self.buf.len())
             .map_err(|e| io::Error::new(io::ErrorKind::InvalidInput, e))?;